menyelesaikan dashboard dan halaman penilaian

+++ routes/web.php
@@ -1,13 +1,19 @@
 <?php
 
+use App\Models\Barang;
+use App\Models\Customer;
+use App\Models\Supplier;
+use App\Charts\KonsumenChart;
+use App\Charts\PenjualanChart;
+use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
+use App\Http\Controllers\HomeController;
+use App\Http\Controllers\SaleController;
+use App\Http\Controllers\OrderController;
 use App\Http\Controllers\BarangController;
 use App\Http\Controllers\GaleriController;
-use App\Http\Controllers\HomeController;
-use App\Http\Controllers\OrderController;
-use App\Http\Controllers\SaleController;
+use App\Http\Controllers\CustomerController;
 use App\Http\Controllers\SupplierController;
-use Illuminate\Support\Facades\Route;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,53 +26,79 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', [AuthController::class, 'index'])->name('login');
-Route::get('/login', [AuthController::class, 'index'])->name('login');
-Route::post('/login', [AuthController::class, 'login']);
-Route::post('/logout', [AuthController::class, 'logout']);
+Route::get('auliamotors/admin', [AuthController::class, 'index'])->name('login');
+Route::get('admin/login', [AuthController::class, 'index'])->name('login');
+Route::post('admin/login', [AuthController::class, 'login']);
 
-Route::get('/dashboard', function () {
-    return view('admin/dashboard');
-})->middleware('auth');
+Route::middleware(['auth'])->group(function () {
+    Route::post('admin/logout', [AuthController::class, 'logout']);
+    Route::get('admin/dashboard', function (PenjualanChart $chart, KonsumenChart $konsumen) {
+        $barang = Barang::all();
+        $supplier = Supplier::all();
+        $customers = Customer::all();
+        $chart =  $chart->build();
+        $konsumen = $konsumen->build();
+        return view('admin/dashboard', compact('chart','konsumen', 'barang', 'supplier', 'customers'));
 
-// Route Barang
-Route::get('/barang', [BarangController::class, 'index'])->middleware('auth');
-Route::post('/barang/store', [BarangController::class, 'store'])->name('barang.store');
-Route::delete('/barang/{barang:kd_barang}', [BarangController::class, 'destroy']);
-Route::put('/barang/edit/{kd_barang}', [BarangController::class, 'update']);
+    });
 
-// Route Supplier
-Route::get('/supplier', [SupplierController::class, 'index']);
-Route::post('/supplier/store', [SupplierController::class, 'store'])->name('supplier.store');
-Route::delete('/supplier/{supplier:kd_supplier}', [SupplierController::class, 'destroy']);
-Route::put('/supplier/edit/{kd_supplier}', [SupplierController::class, 'update']);
+    // Route Barang
+    Route::get('admin/barang', [BarangController::class, 'index'])->middleware('auth');
+    Route::post('admin/barang/store', [BarangController::class, 'store'])->name('barang.store');
+    Route::delete('admin/barang/{barang:kd_barang}', [BarangController::class, 'destroy']);
+    Route::put('admin/barang/edit/{kd_barang}', [BarangController::class, 'update']);
 
-// Pembelian
-Route::get('/pembelian', [OrderController::class, 'index']);
-Route::post('/pembelian/store', [OrderController::class, 'order']);
+    // Route Supplier
+    Route::get('admin/supplier', [SupplierController::class, 'index']);
+    Route::post('admin/supplier/store', [SupplierController::class, 'store'])->name('supplier.store');
+    Route::delete('admin/supplier/{supplier:kd_supplier}', [SupplierController::class, 'destroy']);
+    Route::put('admin/supplier/edit/{kd_supplier}', [SupplierController::class, 'update']);
 
-// Penjualan
-Route::get('/penjualan', [SaleController::class, 'index']);
-Route::post('/penjualan/transaksi', [SaleController::class, 'transaksiJual']);
+    // Pembelian
+    Route::get('admin/pembelian', [OrderController::class, 'index']);
+    Route::post('admin/pembelian/store', [OrderController::class, 'order']);
 
-// Galeri
-Route::get('/galeri', [GaleriController::class, 'index']);
-Route::get('/galeri/{id}', [GaleriController::class, 'show']);
-Route::post('galeri/store', [GaleriController::class, 'store']);
+    // Penjualan
+    Route::get('admin/penjualan', [SaleController::class, 'index']);
+    Route::post('admin/penjualan/transaksi', [SaleController::class, 'transaksiJual']);
 
 
+    // Customer
+    Route::get('admin/customer', [CustomerController::class, 'adminCustomer']);
+
+    // Galeri
+    Route::get('admin/galeri', [GaleriController::class, 'index']);
+    Route::get('admin/galeri/{id}', [GaleriController::class, 'show']);
+    Route::post('admin/galeri/store', [GaleriController::class, 'store']);
+});
+
+
+
+Route::get('/', [HomeController::class, 'index']);
 Route::get('auliamotors', [HomeController::class, 'index']);
 Route::get('auliamotors/services', [HomeController::class, 'services']);
 Route::get('auliamotors/services/motor', [HomeController::class, 'servicesMotor']);
 Route::get('auliamotors/services/steam', [HomeController::class, 'servicesSteam']);
 Route::get('auliamotors/services/modifikasi', [HomeController::class, 'servicesModifikasi']);
 
-
 // pelanggan
-Route::get('auliamotors/customers', function() {
+Route::get('auliamotors/customers', function () {
     return view('customers');
 });
 
-Route::get('auliamotors/contacts', function() {
+// penilaian 
+Route::get('auliamotors/customers/penilaian', function(){
+    return view('grade');
+});
+
+Route::post('auliamotors/customers/penilaian/store', [CustomerController::class, 'store'])->name('penilaian');
+
+// contact
+Route::get('auliamotors/contacts', function () {
     return view('contact');
 });
+
+// galeri
+Route::get('auliamotors/galeri', [HomeController::class, 'galeri']);
+
+
